+++ json_serializable/lib/src/json_key_utils.dart
@@ -209,6 +209,13 @@ KeyConfig _from(FieldElement element, ClassConfig classAnnotation) {
     }
   }
 
+  String? readValueFunctionName;
+  final readValue = obj.read('readValue');
+  if (!readValue.isNull) {
+    final objValue = readValue.objectValue.toFunctionValue()!;
+    readValueFunctionName = objValue.name;
+  }
+
   return _populateJsonKey(
     classAnnotation,
     element,
@@ -217,6 +224,7 @@ KeyConfig _from(FieldElement element, ClassConfig classAnnotation) {
     ignore: obj.read('ignore').literalValue as bool?,
     includeIfNull: obj.read('includeIfNull').literalValue as bool?,
     name: obj.read('name').literalValue as String?,
+    readValueFunctionName: readValueFunctionName,
     required: obj.read('required').literalValue as bool?,
     unknownEnumValue: _annotationValue('unknownEnumValue', mustBeEnum: true),
   );
@@ -230,6 +238,7 @@ KeyConfig _populateJsonKey(
   bool? ignore,
   bool? includeIfNull,
   String? name,
+  String? readValueFunctionName,
   bool? required,
   String? unknownEnumValue,
 }) {
@@ -249,6 +258,7 @@ KeyConfig _populateJsonKey(
     includeIfNull: _includeIfNull(
         includeIfNull, disallowNullValue, classAnnotation.includeIfNull),
     name: name ?? encodedFieldName(classAnnotation.fieldRename, element.name),
+    readValueFunctionName: readValueFunctionName,
     required: required ?? false,
     unknownEnumValue: unknownEnumValue,
   );