+++ json_serializable/test/kitchen_sink/kitchen_sink.g_any_map__checked.dart
@@ -74,6 +74,18 @@ class _Factory implements k.KitchenSinkFactory<dynamic, dynamic> {
       JsonConverterTestClass.fromJson(json);
 }
 
+Object? _valueAccessor(Map json, String key) {
+  if (key == k.trickyKeyName) {
+    return json[k.trickyKeyName] ?? json['STRING'];
+  }
+
+  if (key == 'iterable') {
+    return json['iterable'] ?? json['theIterable'];
+  }
+
+  return json[key];
+}
+
 @JsonSerializable(
   checked: true,
   anyMap: true,
@@ -116,6 +128,7 @@ class KitchenSink implements k.KitchenSink {
 
   BigInt? bigInt;
 
+  @JsonKey(readValue: _valueAccessor)
   Iterable? get iterable => _iterable;
 
   Iterable<dynamic> get dynamicIterable => _dynamicIterable;
@@ -153,7 +166,7 @@ class KitchenSink implements k.KitchenSink {
   // Handle fields with names that collide with helper names
   Map<String, bool> val = _defaultMap();
   bool? writeNotNull;
-  @JsonKey(name: r'$string')
+  @JsonKey(name: k.trickyKeyName, readValue: _valueAccessor)
   String? string;
 
   SimpleObject simpleObject = _defaultSimpleObject();