+++ json_serializable/test/kitchen_sink/kitchen_sink.g_explicit_to_json.g.dart
@@ -10,7 +10,7 @@ part of 'kitchen_sink.g_explicit_to_json.dart';
 
 KitchenSink _$KitchenSinkFromJson(Map<String, dynamic> json) => KitchenSink(
       ctorValidatedNo42: json['no-42'] as int?,
-      iterable: json['iterable'] as List<dynamic>?,
+      iterable: _valueAccessor(json, 'iterable') as List<dynamic>?,
       dynamicIterable: json['dynamicIterable'] as List<dynamic>?,
       objectIterable:
           (json['objectIterable'] as List<dynamic>?)?.map((e) => e as Object),
@@ -80,7 +80,7 @@ KitchenSink _$KitchenSinkFromJson(Map<String, dynamic> json) => KitchenSink(
           .toList()
       ..val = Map<String, bool>.from(json['val'] as Map)
       ..writeNotNull = json['writeNotNull'] as bool?
-      ..string = json[r'$string'] as String?
+      ..string = _valueAccessor(json, r'$string') as String?
       ..simpleObject =
           SimpleObject.fromJson(json['simpleObject'] as Map<String, dynamic>)
       ..strictKeysObject = StrictKeysObject.fromJson(